Basic App
Build a basic android app with java that starts getting the values from light sensor, proximity sensor, accelerometer, and gyroscope, and shows the latest value in 4 different cards in the app screen. Record the sensor values in the android sqlite db every 5 minute, and make time series charts of each of the 4 sensors. If you click on the sensor value cards from the home screen, it goes into a screen with the time series chart corresponding to that sensor. Run the service in the background, so even if the user closes it from the task manager, the app keeps running in the background, it should be showing a notification to let the user know of all the 4 values. The whole app look pretty, be creative.

Reimplement the apis to get data for proximity, accelerometer in android Native Development Kit, increase the data refresh rate from the default rate, with C++ or C and
then use those to get the data and visualize the latest data.